% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nyt_timeseries.R
\name{nyt_timeseries}
\alias{nyt_timeseries}
\title{Collect Articles from The New York Times API}
\usage{
nyt_timeseries(
  api_key,
  query,
  begin_date = "1851-01-01",
  end_date = "2024-01-01",
  continue_loading = FALSE,
  start_page = 0,
  sort = "oldest"
)
}
\arguments{
\item{api_key}{A character string of your New York Times API key.}

\item{query}{A character string specifying the search query for articles.}

\item{begin_date}{A character string specifying the start date for article search in 'YYYY-MM-DD' format. Default: '1851-01-01'.}

\item{end_date}{A character string specifying the end date for article search in 'YYYY-MM-DD' format. Default: '2024-01-01'. Try using Sys.Date() for the current date}

\item{continue_loading}{A logical value indicating whether to continue adding to an existing dataset (TRUE) or start anew (FALSE).}

\item{sort}{A string of characters to determine the order of collecting items. The default is "oldest", it can be changed to "newest".}
}
\value{
A tibble containing articles fetched from the API. This tibble is also stored in the global environment.
}
\description{
This function fetches articles from The New York Times API based on a specified query and date range.
It collects articles page by page and stores the results in a global tibble \code{articles}.
If the function is interrupted, the articles fetched up to that point are retained in \code{articles}.
The function can continue adding to an existing articles dataset or start from scratch.
It also allows specifying a starting page number for continued collection.
}
